backend/requirements.txt
# Backend Dependencies (Core)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-core==2.14.1
pydantic-settings==2.1.0
sqlalchemy>=2.0.25  # Using newer version for Python 3.13 compatibility
aiosqlite==0.19.0
email-validator==2.1.0
python-dotenv==1.0.0

# Security / Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.7

# Rate Limiting & Security
slowapi==0.1.9  # Rate limiting middleware
bleach==6.1.0  # HTML sanitization

# Database Migrations
alembic==1.13.0

# HTTP client / utilities
httpx==0.27.0
faker==20.1.0

# GenAI client
openai==1.3.8

# Note: ML libraries intentionally excluded for initial deployment
# Install as needed: scikit-learn, pandas, numpy, torch, etc.
